refactor: :recycle: Change lib folder into src folder and update references
